fixed .js loader rules for webpack.dev.config so that it compiles the correct files (took off single quotes in 'include' object'). moved server logic to root of repo into server.js and index.js. added dev and heroku-postbuild scripts into package.json. added doctype in index.html. added comments explaining code pieces in more depth.
