--- index.html
@@ -1,9 +1,0 @@
-<html>
-  <head>
-    <meta charset="utf-8">
-  </head>
-  <body>
-    <div id="main"></div>
-    <script type="text/javascript" src="/public/bundle.js"></script>
-  </body>
-</html>
