+++ public/index.html
@@ -1,6 +1,8 @@
-<html>
+<!DOCTYPE html>
+<html lang="en">
   <head>
     <meta charset="utf-8">
+    <title>Webpack React App</title>
   </head>
   <body>
     <div id="main"></div>
