--- src/server.js
@@ -1,15 +1,0 @@
-const path = require('path')
-const express = require('express')
-
-module.exports = {
-  app: function () {
-    const app = express()
-    const indexPath = path.join(__dirname, '../public/index.html')
-    const publicPath = express.static(path.join(__dirname, '../public'))
-
-    app.use('/public', publicPath)
-    app.get('/', function (_, res) { res.sendFile(indexPath) })
-
-    return app
-  }
-}
